pyproject.toml
[tool.poetry]
name = "shift"
version = "0.1.0"
description = ""
authors = ["Rail <rail.akhmietshin@gmail.com>"]
readme = "README.md"

[tool.pytest.ini_options]
pythonpath = [
    ".", "app",
]
asyncio_mode="auto"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
asyncpg = "^0.27.0"
alembic = "^1.11.1"
sqlalchemy = "^2.0.15"
gunicorn = "^20.1.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^11.0.0"}
psycopg2-binary = "^2.9.6"
httpx = "^0.24.1"
Faker = "^18.10.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"



[tool.poetry.dev-dependencies]
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
asyncpg = "^0.27.0"
alembic = "^1.11.1"
SQLAlchemy = "^2.0.15"
gunicorn = "^20.1.0"
fastapi-users = "^11.0.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
